Impove support for 1.4 Firewall Rules Syntax
A big changed happened in `1.4-rolling-202308040557`, and this collection currently cannot handle the syntax after that.  That means one cannot use current rolling releases and expect firewall rules setting to work.

I am certain this is not exhaustive, but it is a big improvement in what actually works!

This partially addresses T6706.

The build number that this change happened is based on information in https://github.com/vyos/vyos-documentation/pull/1076
The regular expression used in this change was tested at https://regex101.com/r/gDIVEI/1